feat: add interface
